refactor(game): decompose JSGame component and extract game logic to custom hook
