+++ src/components/ErrorPage/JSGame/JSGame.component.tsx
@@ -1,61 +1,22 @@
+import { memo } from "react";
 import { motion } from "framer-motion";
-import { memo, useState, useCallback } from "react";
-import { FaCheck, FaTimes } from "react-icons/fa";
-
-type JSQuestion = {
-  id: number;
-  question: string;
-  code: string;
-  options: string[];
-  correctAnswer: number;
-};
-
-const questions: JSQuestion[] = [
-  {
-    id: 1,
-    question: "Co zwróci poniższy kod?",
-    code: "typeof typeof 42",
-    options: ["'number'", "'string'", "'undefined'", "'object'"],
-    correctAnswer: 1
-  },
-  {
-    id: 2,
-    question: "Jaki będzie wynik?",
-    code: "[1, 2, 3].map(num => num * 2)",
-    options: ["[2, 4, 6]", "[1, 2, 3]", "undefined", "Error"],
-    correctAnswer: 0
-  },
-];
-
-type JSGameProps = {
-  onComplete: (score: number) => void;
-};
+import { QuestionHeader } from './components/QuestionHeader.component';
+import { CodeBlock } from './components/CodeBlock.component';
+import { AnswerOption } from './components/AnswerOption.component';
+import { questions } from './constants/questions.data';
+import { JSGameProps } from './types/types';
+import { useGameLogic } from './hooks/useGameLogic.hook';
 
 export const JSGame = memo(({ onComplete }: JSGameProps) => {
-  const [currentQuestion, setCurrentQuestion] = useState(0);
-  const [score, setScore] = useState(0);
-  const [selectedAnswer, setSelectedAnswer] = useState<number | null>(null);
-  const [isCorrect, setIsCorrect] = useState<boolean | null>(null);
+  const {
+    currentQuestion,
+    score,
+    selectedAnswer,
+    isCorrect,
+    handleAnswer
+  } = useGameLogic(onComplete);
 
-  const handleAnswer = useCallback((answerIndex: number) => {
-    setSelectedAnswer(answerIndex);
-    const correct = answerIndex === questions[currentQuestion].correctAnswer;
-    setIsCorrect(correct);
-    
-    if (correct) {
-      setScore(prev => prev + 100);
-    }
-
-    setTimeout(() => {
-      if (currentQuestion < questions.length - 1) {
-        setCurrentQuestion(prev => prev + 1);
-        setSelectedAnswer(null);
-        setIsCorrect(null);
-      } else {
-        onComplete(score + (correct ? 100 : 0));
-      }
-    }, 1500);
-  }, [currentQuestion, onComplete, score]);
+  const currentQuestionData = questions[currentQuestion];
 
   return (
     <motion.div
@@ -65,63 +26,28 @@ export const JSGame = memo(({ onComplete }: JSGameProps) => {
       className="bg-dark/50 backdrop-blur-lg rounded-xl p-6 border border-js/20"
     >
       <div className="mb-6">
-        <div className="flex justify-between items-center mb-4">
-          <span className="text-gray-400">
-            Pytanie {currentQuestion + 1}/{questions.length}
-          </span>
-          <span className="text-js font-bold">
-            Wynik: {score}
-          </span>
-        </div>
+        <QuestionHeader 
+          currentQuestion={currentQuestion}
+          totalQuestions={questions.length}
+          score={score}
+        />
         <h3 className="text-xl text-js mb-4">
-          {questions[currentQuestion].question}
+          {currentQuestionData.question}
         </h3>
-        <pre className="bg-dark/50 p-4 rounded-lg mb-6 overflow-x-auto border border-js/10">
-          <code className="text-js/90 font-mono">
-            {questions[currentQuestion].code}
-          </code>
-        </pre>
+        <CodeBlock code={currentQuestionData.code} />
       </div>
 
       <div className="grid grid-cols-1 gap-4">
-        {questions[currentQuestion].options.map((option, index) => (
-          <motion.button
+        {currentQuestionData.options.map((option, index) => (
+          <AnswerOption
             key={index}
-            whileHover={{ scale: 1.02 }}
-            whileTap={{ scale: 0.98 }}
-            disabled={selectedAnswer !== null}
-            onClick={() => handleAnswer(index)}
-            className={`
-              p-4 rounded-lg text-left transition-all
-              ${selectedAnswer === null 
-                ? 'bg-dark/50 hover:bg-dark/70 text-js border border-js/10' 
-                : selectedAnswer === index
-                  ? isCorrect 
-                    ? 'bg-green-500/20 border-2 border-green-500/50'
-                    : 'bg-red-500/20 border-2 border-red-500/50'
-                  : index === questions[currentQuestion].correctAnswer && !isCorrect
-                    ? 'bg-green-500/20 border-2 border-green-500/50'
-                    : 'bg-dark/50 opacity-50 border border-js/10'
-              }
-            `}
-          >
-            <div className="flex items-center justify-between">
-              <span>{option}</span>
-              {selectedAnswer === index && (
-                <motion.span
-                  initial={{ scale: 0 }}
-                  animate={{ scale: 1 }}
-                  transition={{ type: "spring", stiffness: 200 }}
-                >
-                  {isCorrect ? (
-                    <FaCheck className="text-green-500" />
-                  ) : (
-                    <FaTimes className="text-red-500" />
-                  )}
-                </motion.span>
-              )}
-            </div>
-          </motion.button>
+            option={option}
+            index={index}
+            selectedAnswer={selectedAnswer}
+            isCorrect={isCorrect}
+            correctAnswer={currentQuestionData.correctAnswer}
+            onSelect={handleAnswer}
+          />
         ))}
       </div>
     </motion.div>
