+++ src/components/ErrorPage/JSGame/types/types.ts
@@ -1,0 +1,11 @@
+export type JSQuestion = {
+  id: number;
+  question: string;
+  code: string;
+  options: string[];
+  correctAnswer: number;
+};
+
+export type JSGameProps = {
+  onComplete: (score: number) => void;
+};
